+++ src/config/index.js
@@ -3,3 +3,4 @@ export { REFRESH_TOKEN_EXPIRES_AT } from "./constants";
 export { BASEURL } from "./constants";
 export { default as MESSAGES } from "./messages";
 export { default as ENDPOINTS } from "./endpoints";
+export { default as URLS } from "./urls";
