+++ src/routes/index.js
@@ -1,1 +1,3 @@
-export { default as URLS } from "./Urls";
+export { default as AppRoutes } from "./AppRoutes";
+export { default as PrivateRoute } from "./PrivateRoute";
+export { default as PublicRoute } from "./PublicRoute";
